python-requests/requirements.txt
requests==2.32.3
mysql-connector-python==9.0.0
tabulate==0.9.0
python-dotenv==1.0.1